remove importManyAllVersions, refactor some import methods
