+++ scopes/scope/importer/import-components.ts
@@ -33,6 +33,10 @@ import { FilesStatus, MergeStrategy } from '@teambit/legacy/dist/consumer/versio
 import { MergeResultsThreeWay } from '@teambit/legacy/dist/consumer/versions-ops/merge-version/three-way-merge';
 import ComponentsPendingMerge from '@teambit/legacy/dist/consumer/component-ops/exceptions/components-pending-merge';
 import ManyComponentsWriter from '@teambit/legacy/dist/consumer/component-ops/many-components-writer';
+import ScopeComponentsImporter from '@teambit/legacy/dist/scope/component-ops/scope-components-importer';
+import VersionDependencies, {
+  multipleVersionDependenciesToConsumer,
+} from '@teambit/legacy/dist/scope/version-dependencies';
 
 export type ImportOptions = {
   ids: string[]; // array might be empty
@@ -160,7 +164,7 @@ export default class ImportComponents {
     const bitIds: BitIds = await this.getBitIds();
     const beforeImportVersions = await this._getCurrentVersions(bitIds);
     await this._throwForPotentialIssues(bitIds);
-    const componentsWithDependencies = await this.consumer.importComponentsHarmony(bitIds, true, this.laneObjects);
+    const componentsWithDependencies = await this._importObjectsDisregardLocalCache(bitIds, this.laneObjects);
     if (this.laneObjects && this.options.objectsOnly) {
       // merge the lane objects
       const mergeAllLanesResults = await pMapSeries(this.laneObjects, (laneObject) =>
@@ -226,6 +230,50 @@ export default class ImportComponents {
       );
       return !sameIdHigherVersion;
     });
+  }
+
+  private async _importObjectsDisregardLocalCache(
+    ids: BitIds,
+    lanes: Lane[] = []
+  ): Promise<ComponentWithDependencies[]> {
+    const scopeComponentsImporter = ScopeComponentsImporter.getInstance(this.scope);
+    const versionDependenciesArr: VersionDependencies[] = await scopeComponentsImporter.importMany({
+      ids,
+      cache: false,
+      lanes,
+    });
+    const componentWithDependencies = await pMapSeries(versionDependenciesArr, (versionDependencies) =>
+      versionDependencies.toConsumer(this.scope.objects)
+    );
+    return componentWithDependencies;
+  }
+
+  private async _importComponentsObjects(
+    ids: BitIds,
+    {
+      fromOriginalScope = false,
+      allHistory = false,
+      lane,
+      ignoreMissingHead = false,
+    }: {
+      fromOriginalScope?: boolean;
+      allHistory?: boolean;
+      lane?: Lane;
+      ignoreMissingHead?: boolean;
+    }
+  ): Promise<ComponentWithDependencies[]> {
+    const scopeComponentsImporter = ScopeComponentsImporter.getInstance(this.scope);
+    await scopeComponentsImporter.importManyDeltaWithoutDeps(ids, allHistory, lane, ignoreMissingHead);
+    loader.start(`import ${ids.length} components with their dependencies (if missing)`);
+    const versionDependenciesArr: VersionDependencies[] = fromOriginalScope
+      ? await scopeComponentsImporter.importManyFromOriginalScopes(ids)
+      : await scopeComponentsImporter.importMany({ ids, ignoreMissingHead, lanes: lane ? [lane] : undefined });
+    const componentWithDependencies = await multipleVersionDependenciesToConsumer(
+      versionDependenciesArr,
+      this.scope.objects
+    );
+
+    return componentWithDependencies;
   }
 
   /**
@@ -359,7 +407,7 @@ bit import ${idsFromRemote.map((id) => id.toStringWithoutVersion()).join(' ')}`)
       if (!this.options.objectsOnly) {
         throw new Error(`bit import with no ids and --merge flag was not implemented yet`);
       }
-      componentsAndDependencies = await this.consumer.importComponentsObjects(componentsIdsToImport, {
+      componentsAndDependencies = await this._importComponentsObjects(componentsIdsToImport, {
         fromOriginalScope: this.options.fromOriginalScope,
         allHistory: this.options.allHistory,
       });
